Created separate files (sample and test) for deid_cloud_storage sample
