+++ dlp/snippets/deid_test.py
@@ -404,56 +404,3 @@ def test_deindentify_with_dictionary_replacement(capsys: pytest.CaptureFixture) 
         or "alex@example.com" in out
         or "tal@example.com" in out
     )
-
-
-@mock.patch("google.cloud.dlp_v2.DlpServiceClient")
-def test_deidentify_cloud_storage(
-    dlp_client: MagicMock,
-    capsys: pytest.CaptureFixture,
-) -> None:
-    # Configure the mock DLP client and its behavior.
-    mock_dlp_instance = dlp_client.return_value
-
-    # Configure the mock CreateDlpJob DLP method and its behavior.
-    test_job = f"projects/{GCLOUD_PROJECT}/dlpJobs/test_job"
-    mock_dlp_instance.create_dlp_job.return_value.name = test_job
-
-    # Configure the mock GetDlpJob DLP method and its behavior.
-    mock_job = mock_dlp_instance.get_dlp_job.return_value
-    mock_job.name = test_job
-    mock_job.state = google.cloud.dlp_v2.DlpJob.JobState.DONE
-
-    # Considering this file is present in gcs bucket.
-    file = open(TXT_FILE, "r")
-    # read the content of file
-    data = file.read()
-    # get the length of the data
-    number_of_characters = len(data)
-
-    mock_job.inspect_details.result.processed_bytes = number_of_characters
-    mock_job.inspect_details.result.info_type_stats.info_type.name = "EMAIL_ADDRESS"
-    finding = mock_job.inspect_details.result.info_type_stats.info_type
-
-    mock_job.inspect_details.result.info_type_stats = [
-        MagicMock(info_type=finding, count=1),
-    ]
-
-    deid.deidentify_cloud_storage(
-        GCLOUD_PROJECT,
-        "input_bucket",
-        "output_bucket",
-        ['EMAIL_ADDRESS', 'PERSON_NAME', 'PHONE_NUMBER'],
-        "deidentify_template_name",
-        "structured_deidentify_template_name",
-        "image_redaction_template_name",
-        "DATASET_ID",
-        "TABLE_ID",
-        timeout=1,
-    )
-    out, _ = capsys.readouterr()
-    assert test_job in out
-    assert "Processed Bytes" in out
-    assert "Info type: EMAIL_ADDRESS" in out
-
-    mock_dlp_instance.create_dlp_job.assert_called_once()
-    mock_dlp_instance.get_dlp_job.assert_called_once()
